added OneToMany to Sessions, changed id's to Long instead of Int, checked DTO's and have a proper db migrations folder with a V1 migration

+++ src/main/kotlin/com/scrumdapp/checkpointservice/entities/Checkpoint.kt
@@ -4,7 +4,7 @@ import jakarta.persistence.*
 import java.io.Serializable
 
 @Embeddable
-data class CheckPointId(val groupUserId: Int = 0, val checkpointSessionId: Int = 0) : Serializable
+data class CheckPointId(val groupUserId: Long = 0, val checkpointSessionId: Long = 0) : Serializable
 
 @Entity
 @Table(name = "check_point")
@@ -19,7 +19,7 @@ class Checkpoint {
     @JoinColumn(name = "checkpoint_session_id")
     var checkpointSession: CheckpointSession? = null
 
-    var groupUserId: Int = 0
+    var groupUserId: Long = 0
 
 
     @Column(columnDefinition = "TEXT")
